feat: the folder structure is updated and improved
