+++ Controllers/V1/Category/CategoryControllerDelete.cs
