+++ Controllers/V1/Payments/PaymentControllerDelete.cs
