+++ Controllers/V2/Ordereds/OrderedControllerGet.cs
@@ -8,27 +8,6 @@ namespace RestAdminV2.Controllers
 {
     public partial class OrderedController
     {
-        // GET: api/Customer
-        [HttpGet]
-        public async Task<ActionResult<IEnumerable<Ordered>>> GetOrdereds()
-        {
-            return await _context.Ordereds.Include(i => i.Customer).Include(i => i.Table).ToListAsync();
-        }
-
-        // GET: api/Ordered/5
-        [HttpGet("{id}")]
-        public async Task<ActionResult<Ordered>> GetOrdered(int id)
-        {
-            var ordered = await _context.Ordereds.Include(i => i.Customer).Include(i => i.Table).FirstOrDefaultAsync(i => i.Id == id);
-            if (ordered == null)
-            {
-                return NotFound();
-            }
-
-            return Ok(ordered);
-        }
-
-
         // GET: api/Ordered/TopProducts?month=9
         [HttpGet("TopProducts")]
         public async Task<ActionResult<IEnumerable<Ordered>>> GetTopProducts(int month)
